add comments

+++ assets/script/index.js
@@ -11,6 +11,8 @@ setEmailAndPassword();
  *  Event listeners
 **********************/
 loginButton.addEventListener('click', checkEmailAndPassword);
+
+//If enter is pressed while enter login info
 enterEmail.addEventListener('keydown', (event) => {
     if (event.keyCode === 13) {
         checkEmailAndPassword();
@@ -25,21 +27,22 @@ enterPassword.addEventListener('keydown', (event) => {
  *  Functions
 **********************/
 
-function setEmailAndPassword() {
-    if (localStorage.length < 2) {
+function setEmailAndPassword() { //sets the login and password when user loads page
+    if (localStorage.length < 2) { //if local storage already has the email and password
         const email = 'gabriel@example.com';
         const password = 'notreal';
+
+        //sets email and password
         localStorage.setItem('Email', email);
         localStorage.setItem('Password', password);
-        console.log(localStorage);
     }
 }
 
-function checkEmailAndPassword() {
+function checkEmailAndPassword() { //checks to see if the user entered the correct email and password
     if (enterEmail.value === localStorage.Email && 
-        enterPassword.value === localStorage.Password) {
-            window.open('home.html', '_self');
-        } else {
-            mistake.classList.remove('hidden');
+        enterPassword.value === localStorage.Password) { //if the email and password are correct
+            window.open('home.html', '_self'); //open home page
+        } else { //if the email and password are wrong
+            mistake.classList.remove('hidden'); //shows text informing user that they're wrong
         }
 }
